+++ new-user.html
@@ -98,6 +98,6 @@
         </form>
       </div>
     </main>
-    <script src="/src/js/auth/authentication.mjs" type="module"></script>
+    <script src="/src/js/auth/singup/signup.mjs" type="module"></script>
   </body>
 </html>
