+++ src/js/auth/login/login.mjs
@@ -1,0 +1,5 @@
+import { validateLogin } from "./loginValidation.mjs";
+
+const LOGIN_FORM = document.querySelector("#login-form");
+
+LOGIN_FORM.addEventListener("submit", validateLogin);
