+++ src/js/auth/singup/signup.mjs
@@ -1,0 +1,5 @@
+import { validateSignup } from "./signupValidation.mjs";
+
+const REGISTRATION_FORM = document.querySelector("#registration-form");
+
+REGISTRATION_FORM.addEventListener("submit", validateSignup);
